Initial commit, including calculator, and matrix and fraction modules.
Though this commit is happening in October 2014, this project was begun in
February of that year; hence this initial commit contains a mostly functional
version of the program.
The Fraction class allows for elegant handling of division and rational numbers.
The Matrix class allows for rational-valued matrices of any size.
The calculator itself makes use of these classes, and runs via Reverse Polish
Notation (while also allowing for special commands, for example to open the
help screen or the matrix screen).
Parts of this program do still need to be optimized.
Organization could be useful, splitting the main file into more modules.
Need a Makefile